refactor: doSignUp/doSignIn in api
