+++ functions/authenticator/main_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/dynamodbiface"
 
+	. "github.com/myuon/portals-me/functions/authenticator/api"
 	. "github.com/myuon/portals-me/functions/authenticator/lib"
 	. "github.com/myuon/portals-me/functions/authenticator/verifier"
 	collection "github.com/myuon/portals-me/functions/collection/lib"
